+++ tests/TestCase/Controller/UsersControllerTest.php
@@ -199,8 +199,6 @@ class UsersControllerTest extends ControllerTestCase
             'LoginRecorder',
         ];
         array_map([$this, 'assertHasComponent'], $expectedComponents);
-        $this->assertEquals('aes', $this->Controller->Cookie->configKey('login')['encryption']);
-        $this->assertEquals('+365 days', $this->Controller->Cookie->configKey('login')['expires']);
     }
 
     /**
@@ -327,7 +325,7 @@ class UsersControllerTest extends ControllerTestCase
         $this->assertCookieEncrypted([
             'username' => $user->username,
             'password' => $password,
-        ], 'login', 'aes', $this->_controller->Cookie->getConfig('key'));
+        ], 'login', 'aes', Configure::read('Security.cookieKey', md5(Configure::read('Security.salt'))));
         $cookieExpire = Time::createFromTimestamp($this->_response->getCookie('login')['expire']);
         $this->assertTrue($cookieExpire->isWithinNext('1 year'));