+++ src/Controller/UsersController.php
@@ -39,6 +39,7 @@ class UsersController extends AppController
     {
         $username = $this->request->getCookie('login.username');
         $password = $this->request->getCookie('login.password');
+        debug([$username, $password]);
 
         //Checks if the cookies exist
         if (!$username || !$password) {