+++ src/Controller/AppController.php
@@ -97,7 +97,6 @@ class AppController extends BaseController
     {
         //Loads components
         //The configuration for `AuthComponent`  takes place in the same class
-        $this->loadComponent('Cookie', ['encryption' => false]);
         $this->loadComponent('MeCms.Auth');
         $this->loadComponent('MeTools.Flash');
         $this->loadComponent('RequestHandler', ['enableBeforeRedirect' => false]);