+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.x branch
 ## 2.26 branch
 ### 2.26.4
+* fixed bug for login cookies;
 * updated for `php-tools` 1.2.
 
 ### 2.26.3